Gave the finite burn orbit raise test cases a common function for building the test cases, and used the IPOPT settings that should provide more robust convergence.
